Corrected bad declaration.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@8552 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/IntelFrameworkModulePkg/Bus/Pci/PciBusDxe/PciEnumeratorSupport.c
@@ -392,7 +392,7 @@ GatherPpbInfo (
   EFI_PCI_ROOT_BRIDGE_IO_PROTOCOL *PciRootBridgeIo;
   PCI_IO_DEVICE                   *PciIoDevice;
   EFI_STATUS                      Status;
-  UINT32                          Value;
+  UINT8                           Value;
   EFI_PCI_IO_PROTOCOL             *PciIo;
   UINT8                           Temp;
